Moved Identity.Persistence into Identity.Infrastructure
