+++ CustomCADs.sln
@@ -107,8 +107,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Identity.Domain"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Identity.Application", "src\Modules\Identity\Application\CustomCADs.Identity.Application.csproj", "{5580B332-229C-4C4A-88BF-506E5A7653FC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Identity.Persistence", "src\Modules\Identity\Persistence\CustomCADs.Identity.Persistence.csproj", "{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Identity.Infrastructure", "src\Modules\Identity\Infrastructure\CustomCADs.Identity.Infrastructure.csproj", "{61F88BBB-5E99-4191-8188-9F94ADA4ECDE}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Identity.Endpoints", "src\Modules\Identity\Endpoints\CustomCADs.Identity.Endpoints.csproj", "{63B548BD-A6B8-467C-B823-A1E8FB87DBA0}"
@@ -770,18 +768,6 @@ Global
 		{5580B332-229C-4C4A-88BF-506E5A7653FC}.Release|x64.Build.0 = Release|Any CPU
 		{5580B332-229C-4C4A-88BF-506E5A7653FC}.Release|x86.ActiveCfg = Release|Any CPU
 		{5580B332-229C-4C4A-88BF-506E5A7653FC}.Release|x86.Build.0 = Release|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Debug|x64.Build.0 = Debug|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Debug|x86.Build.0 = Debug|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Release|x64.ActiveCfg = Release|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Release|x64.Build.0 = Release|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Release|x86.ActiveCfg = Release|Any CPU
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4}.Release|x86.Build.0 = Release|Any CPU
 		{61F88BBB-5E99-4191-8188-9F94ADA4ECDE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{61F88BBB-5E99-4191-8188-9F94ADA4ECDE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{61F88BBB-5E99-4191-8188-9F94ADA4ECDE}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1342,7 +1328,6 @@ Global
 		{B8814A68-C612-4FF5-AE28-FFEC8A225A62} = {8CF3001A-FDE9-4E6A-A642-83480DB4B117}
 		{CF02A6A6-4D7C-4229-82BC-CFA77DFBA39B} = {B8814A68-C612-4FF5-AE28-FFEC8A225A62}
 		{5580B332-229C-4C4A-88BF-506E5A7653FC} = {B8814A68-C612-4FF5-AE28-FFEC8A225A62}
-		{9CBDDB42-343E-4DE4-AB4D-450FEC9458E4} = {B8814A68-C612-4FF5-AE28-FFEC8A225A62}
 		{61F88BBB-5E99-4191-8188-9F94ADA4ECDE} = {B8814A68-C612-4FF5-AE28-FFEC8A225A62}
 		{63B548BD-A6B8-467C-B823-A1E8FB87DBA0} = {B8814A68-C612-4FF5-AE28-FFEC8A225A62}
 		{C1CBD8F6-DA83-433C-91B3-4AE8C9DF8A10} = {74636A63-AB16-4950-BF0D-966FBFF79B83}
