+++ src/Modules/Identity/Infrastructure/Identity/Configurations/AppRefreshTokens/Configurations.cs
@@ -1,0 +1,12 @@
+﻿using CustomCADs.Identity.Infrastructure.Identity.ShadowEntities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace CustomCADs.Identity.Infrastructure.Identity.Configurations.AppRefreshTokens;
+
+public class Configurations : IEntityTypeConfiguration<AppRefreshToken>
+{
+	public void Configure(EntityTypeBuilder<AppRefreshToken> builder)
+		=> builder
+			.SetValidations();
+}
