+++ src/Modules/Identity/Infrastructure/Identity/Configurations/AppUserRoles/Configurations.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace CustomCADs.Identity.Infrastructure.Identity.Configurations.AppUserRoles;
+
+using AppUserRole = Microsoft.AspNetCore.Identity.IdentityUserRole<Guid>;
+
+public class Configurations : IEntityTypeConfiguration<AppUserRole>
+{
+	public void Configure(EntityTypeBuilder<AppUserRole> builder)
+		=> builder
+			.SetSeeding();
+}
