+++ src/Modules/Identity/Infrastructure/Identity/IdentityPersistenceReference.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+
+namespace CustomCADs.Identity.Infrastructure.Identity;
+
+public class IdentityPersistenceReference
+{
+	public static Assembly Assembly => typeof(IdentityPersistenceReference).Assembly;
+}
