+++ src/Modules/Identity/Infrastructure/Identity/ShadowEntities/AppRole.cs
@@ -1,0 +1,5 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace CustomCADs.Identity.Infrastructure.Identity.ShadowEntities;
+
+public class AppRole(string name) : IdentityRole<Guid>(name);
