--- src/Modules/Identity/Infrastructure/JwtSettings.cs
@@ -1,8 +1,0 @@
-﻿namespace CustomCADs.Identity.Infrastructure;
-
-public class JwtSettings
-{
-	public required string SecretKey { get; set; }
-	public required string Issuer { get; set; }
-	public required string Audience { get; set; }
-}
