+++ src/Modules/Identity/Infrastructure/Tokens/JwtSettings.cs
@@ -1,0 +1,8 @@
+﻿namespace CustomCADs.Identity.Infrastructure.Tokens;
+
+public class JwtSettings
+{
+	public required string SecretKey { get; set; }
+	public required string Issuer { get; set; }
+	public required string Audience { get; set; }
+}
