--- src/Modules/Identity/Persistence/Configurations/AppRoles/Configurations.cs
@@ -1,12 +1,0 @@
-﻿using CustomCADs.Identity.Persistence.ShadowEntities;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace CustomCADs.Identity.Persistence.Configurations.AppRoles;
-
-public class Configurations : IEntityTypeConfiguration<AppRole>
-{
-	public void Configure(EntityTypeBuilder<AppRole> builder)
-		=> builder
-			.SetSeeding();
-}
