--- src/Modules/Identity/Persistence/IdentityPersistenceReference.cs
@@ -1,8 +1,0 @@
-﻿using System.Reflection;
-
-namespace CustomCADs.Identity.Persistence;
-
-public class IdentityPersistenceReference
-{
-	public static Assembly Assembly => typeof(IdentityPersistenceReference).Assembly;
-}
