--- src/Modules/Identity/Persistence/ShadowEntities/AppRole.cs
@@ -1,5 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace CustomCADs.Identity.Persistence.ShadowEntities;
-
-public class AppRole(string name) : IdentityRole<Guid>(name);
