+++ src/Tools/CodeGen/ProgramExtensions.cs
@@ -2,9 +2,9 @@ using CustomCADs.Customizations.Domain.Services;
 using CustomCADs.Delivery.Infrastructure;
 using CustomCADs.Files.Infrastructure;
 using CustomCADs.Identity.Domain.Users;
-using CustomCADs.Identity.Infrastructure;
-using CustomCADs.Identity.Persistence;
-using CustomCADs.Identity.Persistence.ShadowEntities;
+using CustomCADs.Identity.Infrastructure.Identity;
+using CustomCADs.Identity.Infrastructure.Identity.ShadowEntities;
+using CustomCADs.Identity.Infrastructure.Tokens;
 using CustomCADs.Shared.Infrastructure.Email;
 using CustomCADs.Shared.Infrastructure.Payment;
 using Microsoft.AspNetCore.Identity;
